# docs/swagger.yaml
basePath: /
info:
  contact:
    email: geek@openmind.np
    name: Above&Beyond
  description: API Doc for Suglider Auth
  license:
    name: MIT
    url: https://github.com/tony/suglider-auth/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Suglider-Auth API Doc
  version: "1.0"
paths:
...
  /api/v1/totp/disable:
    put:
      consumes:
      - multipart/form-data
      description: disable TOTP
      parameters:
      - description: Mail
        in: formData
        name: mail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Disable TOTP
      tags:
      - totp
  /api/v1/totp/generate:
    post:
      consumes:
      - multipart/form-data
      description: generate QRcode
      parameters:
      - description: Mail
        in: formData
        name: mail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Enable TOTP
      tags:
      - totp
  /api/v1/totp/validate:
    post:
      consumes:
      - multipart/form-data
      description: If a user has enabled TOTP, the API can be used during the login
        process to verify its validity.
      parameters:
      - description: Mail
        in: formData
        name: mail
        type: string
      - description: OTP Code
        in: formData
        name: otp_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Verify TOTP Validate
      tags:
      - totp
  /api/v1/totp/verify:
    patch:
      consumes:
      - multipart/form-data
      description: The API uses the first enabled TOTP feature to verify the TOTP
        code.
      parameters:
      - description: Mail
        in: formData
        name: mail
        type: string
      - description: OTP Code
        in: formData
        name: otp_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Verify TOTP
      tags:
      - totp
...
  /api/v1/user/mail/disable:
    put:
      consumes:
      - multipart/form-data
      description: Disable Mail OTP feature
      parameters:
      - description: Mail
        in: formData
        name: mail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Mail OTP Disable
      tags:
      - otp
  /api/v1/user/mail/enable:
    put:
      consumes:
      - multipart/form-data
      description: Enable Mail OTP feature
      parameters:
      - description: Mail
        in: formData
        name: mail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Mail OTP Enable
      tags:
      - otp
  /api/v1/user/mail/send:
    post:
      consumes:
      - multipart/form-data
      description: To send an OTP using email.
      parameters:
      - description: Mail
        in: formData
        name: mail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Mail OTP Send
      tags:
      - otp
  /api/v1/user/mail/verify:
    get:
      consumes:
      - multipart/form-data
      description: If a user has enabled Mail OTP, the API can be used during the
        login process to verify its validity.
      parameters:
      - description: Mail
        in: formData
        name: mail
        type: string
      - description: OTP Code
        in: formData
        name: otp_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Mail OTP Verify
      tags:
      - otp
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"